telegram-service/requirements.txt
python-telegram-bot==11.1.0
google-cloud-pubsub==2.17.1